+++ libpcs_io/crc32_sse.c
@@ -222,8 +222,8 @@ uint32_t crc32c(const void *buf, size_t len, uint32_t crc)
 			sse_crc32_u64(crc2, next + LONGx2);
 			next += 8;
 		} while (next < end);
-		crc0 = crc32c_shift(crc32c_long, crc0) ^ crc1;
-		crc0 = crc32c_shift(crc32c_long, crc0) ^ crc2;
+		crc0 = crc32c_shift(crc32c_long, (uint32_t)crc0) ^ crc1;
+		crc0 = crc32c_shift(crc32c_long, (uint32_t)crc0) ^ crc2;
 		next += LONG*2;
 		len -= LONG*3;
 	}
@@ -240,8 +240,8 @@ uint32_t crc32c(const void *buf, size_t len, uint32_t crc)
 			sse_crc32_u64(crc2, next + SHORTx2);
 			next += 8;
 		} while (next < end);
-		crc0 = crc32c_shift(crc32c_short, crc0) ^ crc1;
-		crc0 = crc32c_shift(crc32c_short, crc0) ^ crc2;
+		crc0 = crc32c_shift(crc32c_short, (uint32_t)crc0) ^ crc1;
+		crc0 = crc32c_shift(crc32c_short, (uint32_t)crc0) ^ crc2;
 		next += SHORT*2;
 		len -= SHORT*3;
 	}
